Money and placement requirements, Collision effect
Money is needed to place the turret. Turrent cannot be placed on another one. Some changes to the "Turret" as it was needed for placement issue. Collison of enemies happens now a little bit earlier as it was happening inside of the Planet and I also added some effect during the collision.